Move into workspace-resolver package
