--- packages/dependency-resolver/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../modular-scripts/tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "outDir": "dist-cjs"
-  },
-  "include": ["src/*"],
-  "exclude": ["src/__tests__", "node_modules"]
-}
