+++ packages/workspace-resolver/src/resolve-dependencies/index.ts
